Updated Authentication Fixed some minor bugs
