+++ src/client/layouts/basic-layout/authentication.jsx
@@ -12,6 +12,8 @@ import PropTypes from 'prop-types';
 import app from '../../app';
 import { addNotification } from '../../redux/notifications/actions';
 import { pipe } from '../../../utils/functions';
+import { openDialog } from '../../redux/dialog/actions';
+import { loginUser } from '../../redux/user/actions';
 
 /**
  * Authenticate the user. This also delays rendering the full app so it doesn't flash.
@@ -40,7 +42,7 @@ class Authentication extends PureComponent {
     } catch (error) {
       this.props.addNotification(error.message);
     } finally {
-      await sleep(100);
+      await sleep(200);
 
       this.setState({ hasAuthenticated: true });
     }
@@ -53,10 +55,24 @@ class Authentication extends PureComponent {
     const token = cookie.get('feathers-jwt');
 
     if (token) {
-      await app.authenticate({
+      const { accessToken } = await app.authenticate({
         strategy: 'jwt',
         accessToken: token,
       });
+
+      const verifiedToken = await app.passport.verifyJWT(accessToken);
+
+      cookie.set('feathers-jwt', accessToken);
+
+      app.set('userId', verifiedToken.id);
+
+      const user = await app.service('users').get(verifiedToken.id);
+
+      this.props.loginUser(user);
+
+      if (user.name === null || user.settings.region === null || !user.hasAcceptedTheRules) {
+        this.props.openDialog();
+      }
     }
   };
 
@@ -82,6 +98,14 @@ export default pipe(
     return {
       addNotification(...args) {
         return dispatch(addNotification(...args));
+      },
+
+      loginUser(user) {
+        return dispatch(loginUser(user));
+      },
+
+      openDialog() {
+        return dispatch(openDialog('POST_USER_CREATION_DIALOG'));
       },
     };
   }),
