+++ src/client/layouts/main-layout/main-toolbar/index.js
@@ -1,0 +1,1 @@
+export default from './main-toolbar';
