+++ src/client/redux/user/setup-listeners.js
@@ -1,9 +1,4 @@
-import cookie from 'js-cookie';
-
-import { openDialog } from '../dialog/actions';
-
 import {
-  loginUser,
   logoutUser,
   updateUser,
 } from './actions';
@@ -16,22 +11,6 @@ import {
  */
 export default function setupListeners(app) {
   const users = app.service('users');
-
-  app.on('authenticated', async (payload) => {
-    const verifiedToken = await app.passport.verifyJWT(payload.accessToken);
-
-    cookie.set('feathers-jwt', payload.accessToken);
-
-    app.set('userId', verifiedToken.id);
-
-    const user = await users.get(verifiedToken.id);
-
-    app.store.dispatch(loginUser(user));
-
-    if (user.name === null || user.settings.region === null || !user.hasAcceptedTheRules) {
-      app.store.dispatch(openDialog('POST_USER_CREATION_DIALOG'));
-    }
-  });
 
   users.on('patched', (data) => {
     // Only update when the user is actually online
